added !cc_list, !cc_allow and !cc_deny to the custom commands plugin
!cc_list simply lists all the available custom commands. Only users with the
list_custom_commands permission may do that.

!cc_allow and !cc_deny are aliases to the ACL commands, but their usage can
be granted to people other than owner/op by the owner. That comes in handy
when the owner grants one person the right to configure commands. If that
person uses !cc_set to define an ad-hoc, new command, it's most likely to
be followed by that person wanting to allow $mods to use that new command.
Without !cc_allow, only the owner could update the ACL.
!cc_allow and !cc_deny require the configure_custom_commands_acl permission.